+++ libnopegl/src/darray.h
@@ -27,7 +27,7 @@
 
 struct darray {
     uint8_t *data;
-    int count;
+    size_t count;
     size_t capacity;
     size_t element_size;
     int (*reserve)(struct darray *darray, size_t capacity);
@@ -39,12 +39,12 @@ void *ngli_darray_push(struct darray *darray, const void *element);
 void *ngli_darray_pop(struct darray *darray);
 void *ngli_darray_pop_unsafe(struct darray *darray);
 void *ngli_darray_tail(const struct darray *darray);
-void *ngli_darray_get(const struct darray *darray, int index);
-void ngli_darray_remove(struct darray *darray, int index);
-void ngli_darray_remove_range(struct darray *darray, int index, int count);
+void *ngli_darray_get(const struct darray *darray, size_t index);
+void ngli_darray_remove(struct darray *darray, size_t index);
+void ngli_darray_remove_range(struct darray *darray, size_t index, size_t count);
 void ngli_darray_reset(struct darray *darray);
 
-static inline int ngli_darray_count(const struct darray *darray)
+static inline size_t ngli_darray_count(const struct darray *darray)
 {
     return darray->count;
 }
